:truck: separate types & index for export on textarea
